requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
sentence-transformers
pypdf
python-dotenv
reportlab
deep-translator
fastapi
uvicorn
python-multipart
langchain-groq
requests
beautifulsoup4
lxml
gunicorn
